snies-backend/requirements.docker.txt
asgiref==3.8.1
Django==4.2.15
sqlparse==0.5.1
tzdata==2024.1
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
gunicorn==21.2.0
django-cors-headers==4.6.0
openpyxl==3.1.5
